feat: member 통합 entity , dto 값 검증 의존성 추가
