+++ .idea/modules.xml
@@ -5,7 +5,6 @@
       <module fileurl="file://$PROJECT_DIR$/.idea/SwithT_BE.iml" filepath="$PROJECT_DIR$/.idea/SwithT_BE.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/lecture.main.iml" filepath="$PROJECT_DIR$/.idea/modules/lecture.main.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/lesson.main.iml" filepath="$PROJECT_DIR$/.idea/modules/lesson.main.iml" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/member.main.iml" filepath="$PROJECT_DIR$/.idea/modules/member.main.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/payment.main.iml" filepath="$PROJECT_DIR$/.idea/modules/payment.main.iml" />
     </modules>
   </component>
