--- member/src/main/java/com/tweety/SwithT/common/domain/Role.java
@@ -1,6 +1,0 @@
-package com.tweety.SwithT.common.domain;
-
-public enum Role {
-    MAN,WOMAN;
-}
-
