+++ member/src/main/java/com/tweety/SwithT/member/controller/MemberController.java
@@ -1,0 +1,17 @@
+package com.tweety.SwithT.member.controller;
+
+import com.tweety.SwithT.member.service.MemberService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class MemberController {
+
+    private final MemberService memberService;
+
+    @Autowired
+    public MemberController(MemberService memberService) {
+        this.memberService = memberService;
+    }
+
+}
