+++ member/src/main/java/com/tweety/SwithT/member/domain/Gender.java
@@ -1,0 +1,6 @@
+package com.tweety.SwithT.member.domain;
+
+public enum Gender {
+    MAN,WOMAN;
+}
+
