+++ member/src/main/java/com/tweety/SwithT/member/domain/Role.java
@@ -1,0 +1,8 @@
+package com.tweety.SwithT.member.domain;
+
+public enum Role {
+    ADMIN,
+    TUTOR,
+    TUTEE;
+}
+
