+++ member/src/main/java/com/tweety/SwithT/member/dto/MemberSaveReqDto.java
@@ -1,14 +1,17 @@
-package com.tweety.SwithT.tutor.dto;
+package com.tweety.SwithT.member.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+@Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-@Data
-public class TutorDeleteDto {
+public class MemberSaveReqDto {
+
+
+
 
 }
