+++ member/src/main/java/com/tweety/SwithT/member/repository/MemberRepository.java
@@ -1,0 +1,12 @@
+package com.tweety.SwithT.member.repository;
+
+import com.tweety.SwithT.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+
+
+}
