+++ member/src/main/java/com/tweety/SwithT/member/service/MemberService.java
@@ -1,0 +1,18 @@
+package com.tweety.SwithT.member.service;
+
+import com.tweety.SwithT.member.repository.MemberRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Transactional
+@Service
+public class MemberService {
+
+    private final MemberRepository memberRepository;
+
+    @Autowired
+    public MemberService(MemberRepository memberRepository) {
+        this.memberRepository = memberRepository;
+    }
+}
