--- member/src/main/java/com/tweety/SwithT/tutee/dto/TuteeListResDto.java
@@ -1,14 +1,0 @@
-package com.tweety.SwithT.tutee.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class TuteeListResDto {
-
-}
