--- member/src/main/java/com/tweety/SwithT/tutor/dto/TutorListResDto.java
@@ -1,14 +1,0 @@
-package com.tweety.SwithT.tutor.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class TutorListResDto {
-
-}
