--- member/src/main/java/com/tweety/SwithT/tutor/service/TutorService.java
@@ -1,17 +1,0 @@
-package com.tweety.SwithT.tutor.service;
-
-import com.tweety.SwithT.tutor.repository.TutorRepository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Transactional
-@Service
-public class TutorService {
-
-    private final TutorRepository tutorRepository;
-
-    public TutorService(TutorRepository tutorRepository) {
-        this.tutorRepository = tutorRepository;
-    }
-
-}
